+++ hutool-core/src/main/java/cn/hutool/core/stream/CollectorUtil.java
@@ -14,7 +14,6 @@ import java.util.function.BiConsumer;
 import java.util.function.BinaryOperator;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.function.UnaryOperator;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;