+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * 【core   】     修复ChineseDate在1970年之前农历差一天问题（issue#I4UTPK@Gitee）
 * 【core   】     修复CoordinateUtil精准问题及转换bug（pr#551@Gitee）
 * 【json   】     修复JSONObject解析XML后没有返回的bug（issue#2160@Github）
+* 【extra  】     修复GanymedUtil错误信息读取位置问题（issue#I4VDZ2@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.7.21 (2022-02-14)